requirements.txt
requests
psycopg2-binary
gspread
oauth2client
pandas
python-dotenv